refactor(rust): extract worktree.rs with git worktree commands
